# openapi.yaml
openapi: 3.0.3
info:
  title: Content Service API (Social Network)
  version: 1.0.0
  description: REST API для управления постами в микроблоге.
paths:
  /posts:
    get:
      summary: Получить список постов
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Номер страницы
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Количество элементов на странице
        - in: query
          name: clan_emoji
          schema:
            type: string
          description: Фильтр по эмодзи клана
      responses:
        '200':
          description: Список постов
    post:
      summary: Создать пост
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [author_id, clan_emoji, content]
              properties:
                author_id:
                  type: string
                clan_emoji:
                  type: string
                content:
                  type: string
                  minLength: 1
                  maxLength: 280
      responses:
        '201':
          description: Пост создан
        '400':
          $ref: '#/components/responses/ValidationError'
  /posts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    get:
      summary: Получить пост по ID
      responses:
        '200':
          description: Данные поста
        '404':
          description: Пост не найден
    put:
      summary: Обновить пост
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 280
      responses:
        '200':
          description: Пост обновлен
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Пост не найден
    delete:
      summary: Удалить пост
      responses:
        '204':
          description: Пост успешно удален
        '404':
          description: Пост не найден

components:
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Validation Failed"
              details:
                type: array
                items:
                  type: string
                example: ["content must be between 1 and 280 characters"]